Improve logging & errors; Cleanup
Improve logging
Improve error handling
Make DB singleton with auto teardown
Cleanup typed pickers
Give typed pickers prios
Extend lcsc with symbol info for heuristic attach
Fix paths
