+++ src/faebryk/library/has_multi_picker.py
@@ -20,9 +20,8 @@ class has_multi_picker(has_picker.impl()):
                 picker.pick(self.get_obj())
                 return
             except PickError as e:
-                logger.info(f"Picker {picker} for module {self.get_obj()} failed: {e}")
+                logger.debug(f"Picker {picker} for module {self.get_obj()} failed: {e}")
         raise LookupError("All pickers failed")
-
 
     class Picker:
         @abstractmethod
